@@ -1,2 +1,1 @@
-# version of the climate_indices package, should match with setup.py
-__version__ = "1.0.10"
+"""Climate indices for drought monitoring"""
