@@ -1,104 +1,33 @@
+"""Various utility/convenience functions"""
+
 import calendar
 from datetime import datetime
 import logging
 
-import numba
 import numpy as np
 
-
-# ------------------------------------------------------------------------------
-# set up a basic, global _logger
-def get_logger(name, level):
-    logging.basicConfig(
-        format="%(asctime)s %(levelname)s %(message)s",
-        datefmt="%Y-%m-%d  %H:%M:%S",
-    )
-    logger = logging.getLogger(name)
-    logger.setLevel(level)
-    return logger
-
-
-# ------------------------------------------------------------------------------
-def sign_change(
-        a: np.ndarray,
-        b: np.ndarray,
-) -> np.ndarray:
-    """
-    Given two same-sized arrays of floats return an array of booleans indicating
-    if a sign change occurs at the corresponding index.
-
-    :param a: array of floats
-    :param b: array of floats
-    :return: array of booleans of same size as input arrays
-    """
-
-    if a.size != b.size:
-        raise ValueError("Mismatched input arrays")
-
-    # use the shape of the first array as the shape of the array we'll return
-    original_shape = a.shape
-
-    # get the sign value for each element
-    sign_a = np.sign(a.flatten())
-    sign_b = np.sign(b.flatten())
-
-    # sign change between the two where values unequal
-    sign_changes = sign_a != sign_b
-
-    return np.reshape(sign_changes, original_shape)
+# declare the function names that should be included in the public API for this module
+__all__ = [
+    "compute_days",
+    "count_zeros_and_non_missings",
+    "get_logger",
+    "get_tolerance",
+    "gregorian_length_as_366day",
+    "is_data_valid",
+    "reshape_to_2d",
+    "reshape_to_divs_years_months",
+    "rmse",
+    "sign_change",
+    "transform_to_366day",
+    "transform_to_gregorian",
+]
 
 
-# ------------------------------------------------------------------------------
-def is_data_valid(
-        data: np.ndarray,
-) -> bool:
-    """
-    Returns if an array is valid or not, i.e. a supported array type
-    (ndarray or MaskArray) which is not all-NaN.
-
-    :param data: data object, expected as either numpy.ndarry or numpy.ma.MaskArray
-    :return True if array is non-NaN for at least one element
-        and is an array type valid for processing by other modules
-    :rtype: boolean
-    """
-
-    # make sure we're not dealing with all NaN values
-    if np.ma.isMaskedArray(data):
-
-        valid_flag = bool(data.count())
-
-    elif isinstance(data, np.ndarray):
-
-        valid_flag = not np.all(np.isnan(data))
-
-    else:
-        _logger.warning("Invalid data type")
-        valid_flag = False
-
-    return valid_flag
-
-
-# ------------------------------------------------------------------------------
-def rmse(
-        predictions: np.ndarray,
-        targets: np.ndarray,
-) -> np.ndarray:
-    """
-    Root mean square error
-
-    :param predictions: np.ndarray
-    :param targets: np.ndarray
-    :return: np.ndarray
-    """
-    return np.sqrt(((predictions - targets) ** 2).mean())
-
-
-# ------------------------------------------------------------------------------
 def compute_days(
-        initial_year: int,
-        total_months: int,
-        initial_month=1,
-        units_start_year=1800,
+    initial_year: int,
+    total_months: int,
+    initial_month=1,
+    units_start_year=1800,
 ) -> np.ndarray:
     """
     Computes the "number of days" equivalent for regular, incremental monthly
@@ -129,10 +58,7 @@ def compute_days(
 
     # loop over all time steps (months)
     for i in range(total_months):
-
-        years = int(
-            (i + initial_month - 1) / 12
-        )  # the number of years since the initial year
+        years = int((i + initial_month - 1) / 12)  # the number of years since the initial year
         months = int((i + initial_month - 1) % 12)  # the number of months since January
 
         # cook up a datetime object for the current time step (month)
@@ -144,11 +70,129 @@ def compute_days(
     return days
 
 
-# ------------------------------------------------------------------------------
-@numba.jit
+def count_zeros_and_non_missings(
+    values: np.ndarray,
+) -> (int, int):
+    """
+    Given an input array of values return a count of the zeros
+    and non-missing values. Missing values assumed to be numpy.NaNs.
+
+    :param values: array like object (numpy array, most likely)
+    :return: two int scalars: 1) the count of zeros, and
+        2) the count of non-missing values
+    """
+
+    # make sure we have a numpy array
+    values = np.array(values)
+
+    # count the number of zeros and non-missing (non-NaN) values
+    zeros = values.size - np.count_nonzero(values)
+    non_missings = np.count_nonzero(~np.isnan(values))
+
+    return zeros, non_missings
+
+
+def get_logger(name, level):
+    """
+    Sets up a basic, global _logger
+
+    :param name:
+    :param level:
+    :return:
+    """
+    logging.basicConfig(
+        format="%(asctime)s %(levelname)s %(message)s",
+        datefmt="%Y-%m-%d  %H:%M:%S",
+    )
+    logger = logging.getLogger(name)
+    logger.setLevel(level)
+    return logger
+
+
+def get_tolerance(dim: np.ndarray) -> float:
+    """
+    dynamic threshold absolute tolerance parameter np.allclose
+    derived from (smallest) absolute grid size along dimension dim.
+    Always greater than zero.
+    """
+    tol = np.abs(np.diff(dim)).min() / 10
+    return max(tol, np.finfo(tol.dtype).resolution)
+
+
+def sign_change(
+    a: np.ndarray,
+    b: np.ndarray,
+) -> np.ndarray:
+    """
+    Given two same-sized arrays of floats return an array of booleans indicating
+    if a sign change occurs at the corresponding index.
+
+    :param a: array of floats
+    :param b: array of floats
+    :return: array of booleans of same size as input arrays
+    """
+
+    if a.size != b.size:
+        raise ValueError("Mismatched input arrays")
+
+    # use the shape of the first array as the shape of the array we'll return
+    original_shape = a.shape
+
+    # get the sign value for each element
+    sign_a = np.sign(a.flatten())
+    sign_b = np.sign(b.flatten())
+
+    # sign change between the two where values unequal
+    sign_changes = sign_a != sign_b
+
+    return np.reshape(sign_changes, original_shape)
+
+
+def is_data_valid(
+    data: np.ndarray,
+) -> bool:
+    """
+    Returns if an array is valid or not, i.e. a supported array type
+    (ndarray or MaskArray) which is not all-NaN.
+
+    :param data: data object, expected as either numpy.ndarry or numpy.ma.MaskArray
+    :return True if array is non-NaN for at least one element
+        and is an array type valid for processing by other modules
+    :rtype: boolean
+    """
+
+    # make sure we're not dealing with all NaN values
+    if np.ma.isMaskedArray(data):
+        # TODO fix this, there is no ndarray.count according to PyCharm's warning, use another approach for this flag
+        valid_flag = bool(data.count())
+
+    elif isinstance(data, np.ndarray):
+        valid_flag = not np.all(np.isnan(data))
+
+    else:
+        _logger.warning("Invalid data type")
+        valid_flag = False
+
+    return valid_flag
+
+
+def rmse(
+    predictions: np.ndarray,
+    targets: np.ndarray,
+) -> np.ndarray:
+    """
+    Root mean square error
+
+    :param predictions: np.ndarray
+    :param targets: np.ndarray
+    :return: np.ndarray
+    """
+    return np.sqrt(((predictions - targets) ** 2).mean())
+
+
 def reshape_to_2d(
-        values: np.ndarray,
-        second_axis_length: int,
+    values: np.ndarray,
+    second_axis_length: int,
 ) -> np.ndarray:
     """
     :param values: an 1-D numpy.ndarray of values
@@ -165,8 +209,9 @@ def reshape_to_2d(
             # data is already in the shape we want, return it unaltered
             return values
         else:
-            message = "Values array has an invalid shape (2-D but second " + \
-                      f"dimension not {second_axis_length}): {shape}"
+            message = (
+                "Values array has an invalid shape (2-D but second " + f"dimension not {second_axis_length}): {shape}"
+            )
             _logger.error(message)
             raise ValueError(message)
 
@@ -181,10 +226,7 @@ def reshape_to_2d(
     final_year_values = shape[0] % second_axis_length
     if final_year_values > 0:
         pads = second_axis_length - final_year_values
-        values = np.pad(values,
-                        pad_width=(0, pads),
-                        mode="constant",
-                        constant_values=np.NaN)
+        values = np.pad(values, pad_width=(0, pads), mode="constant", constant_values=np.NaN)
 
     # we should have an ordinal number of years now
     # (ordinally divisible by second_axis_length)
@@ -194,10 +236,8 @@ def reshape_to_2d(
     return np.reshape(values, newshape=(first_axis_length, second_axis_length))
 
 
-# ------------------------------------------------------------------------------
-@numba.jit
 def reshape_to_divs_years_months(
-        monthly_values: np.ndarray,
+    monthly_values: np.ndarray,
 ) -> np.ndarray:
     """
     :param monthly_values: an 2-D numpy.ndarray of monthly values,
@@ -216,8 +256,7 @@ def reshape_to_divs_years_months(
             # data is already in the shape we want, return it unaltered
             return monthly_values
         else:
-            message = "Values array has an invalid shape (3-D but " + \
-                      "third dimension is not 12): " + str(shape)
+            message = "Values array has an invalid shape (3-D but " + "third dimension is not 12): " + str(shape)
             _logger.error(message)
             raise ValueError(message)
 
@@ -230,8 +269,7 @@ def reshape_to_divs_years_months(
     # otherwise make sure that we've been passed in a 2-D array
     # of values with the final dimension size == 12
     elif shape[1] != 12:
-        message = "Values array has an invalid shape (second dimension " + \
-                  "should be 12, but is not): " + str(shape)
+        message = "Values array has an invalid shape (second dimension " + "should be 12, but is not): " + str(shape)
         _logger.error(message)
         raise ValueError(message)
 
@@ -243,18 +281,14 @@ def reshape_to_divs_years_months(
     return np.reshape(monthly_values, (shape[0], total_years, 12))
 
 
-# ------------------------------------------------------------------------------
-@numba.jit
 def gregorian_length_as_366day(
-        length_gregorian: int,
-        year_start: int,
+    length_gregorian: int,
+    year_start: int,
 ) -> int:
-
     year = year_start
     remaining = length_gregorian
     length_366day = 0
     while remaining > 0:
-
         if calendar.isleap(year):
             days_in_current_year = 366
         else:
@@ -271,12 +305,10 @@ def gregorian_length_as_366day(
     return length_366day
 
 
-# ------------------------------------------------------------------------------
-@numba.jit
 def transform_to_366day(
-        original: np.ndarray,
-        year_start: int,
-        total_years: int,
+    original: np.ndarray,
+    year_start: int,
+    total_years: int,
 ) -> np.ndarray:
     """
     Takes an array of daily values with only actual leap years represented
@@ -296,7 +328,7 @@ def transform_to_366day(
 
     :param original: 1-D array of daily values
     :param year_start: the year corresponding to the initial year of the input
-        array, used to determine whether or not each increment of daily values
+        array, used to determine whether  each increment of daily values
         represents an actual leap year
     :param total_years: the total number of years represented by the input array
     :return: 1-D array of values with size (total_years * 366)
@@ -320,28 +352,22 @@ def transform_to_366day(
 
     # loop over each year
     for year in range(year_start, year_start + total_years):
-
         if calendar.isleap(year):
-
             # write the next 366 days from the original time
             # series into the all_leap array
-            all_leap[all_leap_index:(all_leap_index + 366)] = \
-                original[original_index:(original_index + 366)]
+            all_leap[all_leap_index: (all_leap_index + 366)] = original[original_index: (original_index + 366)]
 
             # increment the "start day of the current year" index for the original
             # so that the next iteration jumps ahead a full year
             original_index += 366
 
         else:
-
             # write the first 59 days (Jan 1 through Feb 28) from
             # the original time series into the all_leap array
-            all_leap[all_leap_index:(all_leap_index + 59)] = \
-                original[original_index:(original_index + 59)]
+            all_leap[all_leap_index: (all_leap_index + 59)] = original[original_index: (original_index + 59)]
 
             # average the Feb 28th and March 1st values as the faux Feb 29th value
-            all_leap[all_leap_index + 59] = \
-                (original[original_index + 58] + original[original_index + 59]) / 2
+            all_leap[all_leap_index + 59] = (original[original_index + 58] + original[original_index + 59]) / 2
 
             # write the remaining days of the year (Mar 1 through Dec 31)
             # from the original into the all_leap array
@@ -351,7 +377,15 @@ def transform_to_366day(
                 remainder = original[original_index + 59:]
                 difference = len(all_leap[all_leap_index + 60:]) - len(remainder)
                 if difference > 0:
-                    final_days = np.pad(remainder, (0, difference,), mode='constant', constant_values=np.NaN)
+                    final_days = np.pad(
+                        remainder,
+                        (
+                            0,
+                            difference,
+                        ),
+                        mode="constant",
+                        constant_values=np.NaN,
+                    )
                 elif difference != 0:
                     raise ValueError("Incompatible shapes")
                 else:
@@ -359,8 +393,9 @@ def transform_to_366day(
                 all_leap[all_leap_index + 60:] = final_days
                 continue
             else:
-                all_leap[all_leap_index + 60:(all_leap_index + 366)] = \
-                    original[original_index + 59:original_year_end_index]
+                all_leap[all_leap_index + 60: (all_leap_index + 366)] = original[
+                    original_index + 59: original_year_end_index
+                ]
 
             # increment the "start day of the current year" index for the original
             # so the next iteration jumps ahead a full year
@@ -371,11 +406,9 @@ def transform_to_366day(
     return all_leap
 
 
-# ------------------------------------------------------------------------------
-@numba.jit
 def transform_to_gregorian(
-        original: np.ndarray,
-        year_start: int,
+    original: np.ndarray,
+    year_start: int,
 ) -> np.ndarray:
     """
     Takes an array of daily values represented as full 366-day years (as if each
@@ -400,7 +433,7 @@ def transform_to_gregorian(
     :param original: 1-D array of daily values, total size should be
         a multiple of 366
     :param year_start: the year corresponding to the initial year (first 366
-        values) of the input array, used to determine whether or not each 366
+        values) of the input array, used to determine whether each 366
         increment of daily values represents an actual leap year
     """
     # original time series is assumed to be a one-dimensional array of floats
@@ -413,8 +446,7 @@ def transform_to_gregorian(
         _logger.error(message)
         raise ValueError(message)
     if original.size % 366 != 0:
-        message = "Invalid input array: only 1-D arrays containing " + \
-                  "multiples of 366 days are supported"
+        message = "Invalid input array: only 1-D arrays containing " + "multiples of 366 days are supported"
         _logger.error(message)
         raise ValueError(message)
 
@@ -433,29 +465,25 @@ def transform_to_gregorian(
 
     # loop over each year
     for year in range(year_start, year_start + total_years):
-
         if calendar.isleap(year):
-
             # write the next 366 days from the original
             # time series into the gregorian array
-            gregorian[gregorian_index:(gregorian_index + 366)] = \
-                original[original_index:(original_index + 366)]
+            gregorian[gregorian_index: (gregorian_index + 366)] = original[original_index: (original_index + 366)]
 
             # increment the "start day of the current year" index for the original
             # so the next iteration jumps ahead a full year
             gregorian_index += 366
 
         else:
-
             # write the first 59 days (Jan 1 through Feb 28) from the original
             # time series into the gregorian array
-            gregorian[gregorian_index:(gregorian_index + 59)] = \
-                original[original_index:(original_index + 59)]
+            gregorian[gregorian_index: (gregorian_index + 59)] = original[original_index: (original_index + 59)]
 
             # write the remaining days of the year (Mar 1 through Dec 31)
             # from the original into the gregorian array
-            gregorian[(gregorian_index + 59):(gregorian_index + 365)] = \
-                original[(original_index + 60):(original_index + 366)]
+            gregorian[(gregorian_index + 59): (gregorian_index + 365)] = original[
+                (original_index + 60): (original_index + 366)
+            ]
 
             # increment the "start day of the current year" index for
             # the original so the next iteration jumps ahead a full year
@@ -464,40 +492,6 @@ def transform_to_gregorian(
         original_index += 366
 
     return gregorian
-
-
-# ------------------------------------------------------------------------------
-def count_zeros_and_non_missings(
-        values: np.ndarray,
-) -> (int, int):
-    """
-    Given an input array of values return a count of the zeros
-    and non-missing values. Missing values assumed to be numpy.NaNs.
-
-    :param values: array like object (numpy array, most likely)
-    :return: two int scalars: 1) the count of zeros, and
-        2) the count of non-missing values
-    """
-
-    # make sure we have a numpy array
-    values = np.array(values)
-
-    # count the number of zeros and non-missing (non-NaN) values
-    zeros = values.size - np.count_nonzero(values)
-    non_missings = np.count_nonzero(~np.isnan(values))
-
-    return zeros, non_missings
-
-
-# ------------------------------------------------------------------------------
-def get_tolerance(dim: np.ndarray) -> float:
-    """
-    dynamic threshold absolute tolerance parameter np.allclose
-    derived from (smallest) absolute grid size along dimension dim.
-    Always greater than zero.
-    """
-    tol = np.abs(np.diff(dim)).min() / 10
-    return max(tol, np.finfo(tol.dtype).resolution)
 
 
 _logger = get_logger(__name__, logging.DEBUG)
