@@ -84,49 +84,23 @@ def latitude_degrees():
 
 
 @pytest.fixture(scope="module")
-def awc_inches():
-    return _AWC_INCHES
-
-
-@pytest.fixture(scope="module")
-def palmer_awc():
-    return _AWC_INCHES_PALMER
-
-
-@pytest.fixture(scope="module")
 def precips_mm_monthly():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "precips_mm_monthly.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "precips_mm_monthly.npy"))
 
 
 @pytest.fixture(scope="module")
 def precips_mm_daily():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "precips_mm_daily.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "precips_mm_daily.npy"))
 
 
 @pytest.fixture(scope="module")
 def transformed_pearson3():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "pearson3_monthly.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "pearson3_monthly.npy"))
 
 
 @pytest.fixture(scope="module")
 def transformed_pearson3_monthly_fullperiod():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "pearson3_monthly_full.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "pearson3_monthly_full.npy"))
 
 
 @pytest.fixture(scope="module")
@@ -138,10 +112,7 @@ def gamma_monthly():
 
     :return:
     """
-    return np.load(
-        os.path.join(os.path.split(__file__)[0], "fixture",
-                     "gamma_monthly.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "gamma_monthly.npy"))
 
 
 @pytest.fixture(scope="module")
@@ -154,11 +125,7 @@ def gamma_daily():
     :return:
     """
 
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "gamma_daily.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "gamma_daily.npy"))
 
 
 @pytest.fixture(scope="module")
@@ -170,10 +137,7 @@ def transformed_gamma_monthly():
 
     :return:
     """
-    return np.load(
-        os.path.join(os.path.split(__file__)[0], "fixture",
-                     "transformed_gamma_monthly.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "transformed_gamma_monthly.npy"))
 
 
 @pytest.fixture(scope="module")
@@ -186,325 +150,59 @@ def transformed_gamma_daily():
     :return:
     """
 
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "transformed_gamma_daily.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "transformed_gamma_daily.npy"))
 
 
 @pytest.fixture(scope="module")
 def pet_thornthwaite_mm():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "pet_thornthwaite.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "pet_thornthwaite.npy"))
 
 
 @pytest.fixture(scope="module")
 def temps_celsius():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "temp_celsius.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "temp_celsius.npy"))
 
 
 @pytest.fixture(scope="module")
 def pnp_6month():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "pnp_06.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "pnp_06.npy"))
 
 
 @pytest.fixture(scope="module")
 def spei_6_month_gamma():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "spei_06_gamma.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "spei_06_gamma.npy"))
 
 
 @pytest.fixture(scope="module")
 def spei_6_month_pearson3():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "spei_06_pearson3.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "spei_06_pearson3.npy"))
 
 
 @pytest.fixture(scope="module")
 def spi_1_month_gamma():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "spi_01_gamma.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "spi_01_gamma.npy"))
 
 
 @pytest.fixture(scope="module")
 def spi_6_month_gamma():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "spi_06_gamma.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "spi_06_gamma.npy"))
 
 
 @pytest.fixture(scope="module")
 def spi_6_month_pearson3():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "spi_06_pearson3.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "spi_06_pearson3.npy"))
 
 
-@pytest.fixture(scope="module")
-def palmer_pet():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_pet.npy")
-    )
+@pytest.fixture(scope="function")
+def rain_mm() -> np.ndarray:
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "rain_mm.npy"))
 
 
-@pytest.fixture(scope="module")
-def palmer_precip():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_precip.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_et():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_et.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_pr():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_pr.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_r():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_r.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_ro():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_ro.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_pro():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_pro.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_l():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_l.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_pl():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_pl.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_pdsi_monthly():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_pdsi.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_pdsi_from_scpdsi_monthly():
-    return np.load(
-        os.path.join(
-            os.path.split(__file__)[0],
-            "fixture",
-            "palmer_pdsi_from_scpdsi.npy"
-        )
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_phdi_monthly():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_phdi.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_pmdi_monthly():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_pmdi.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_zindex():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_zindex_0.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_alpha():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_alpha.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_beta():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_beta.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_gamma():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_gamma.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_delta():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_delta.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_K():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_k.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_zindex_monthly():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_zindex_1.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_scpdsi_monthly():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_scpdsi.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_scphdi_monthly():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_scphdi.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_scpmdi_monthly():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_scpmdi.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def palmer_sczindex_monthly():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "palmer_sczindex.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def rain_mm():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "rain_mm.npy")
-    )
-
-
-@pytest.fixture(scope="module")
-def rain_mm_365():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "rain_mm_365.npy")
-    )
+@pytest.fixture(scope="function")
+def rain_mm_365() -> np.ndarray:
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "rain_mm_365.npy"))
 
 
 @pytest.fixture(scope="module")
 def rain_mm_366():
-    return np.load(
-        os.path.join(os.path.split(__file__)[0],
-                     "fixture",
-                     "rain_mm_366.npy")
-    )
+    return np.load(os.path.join(os.path.split(__file__)[0], "fixture", "rain_mm_366.npy"))
